requirements.txt
python-telegram-bot
python-dotenv
telethon
pandas
pyarrow
matplotlib
pillow
pytest
pytest-mock
pytesseract
opencv-python